nx file io package to primitive
